[velox] refactor(cxl): Move non-CXL node check into the allocator
The check that rejects a CPU-bearing (non-CXL) NUMA node moves from CxlMemoryResource::create into CxlMemoryAllocator's constructor, inside validateNumaNode alongside the existing availability and range checks. The allocator is what binds pages to the node, so it is where a silent DRAM misbinding would happen. Enforcing the invariant there closes the hole where a directly-constructed allocator skipped the check, and CxlMemoryResource::create no longer needs the NUMA includes or the TestValue hook.

The allocator's node-0 tests become DEBUG_ONLY_TEST_F and inject the TestValue hook to treat the bound node as CXL, so they remain honest about running on a DRAM host. A new rejectsNonCxlNode test binds the node hosting CPU 0 and asserts the failure in every build mode.
