+++ velox/experimental/cxl/CxlMemoryAllocator.cpp
@@ -18,8 +18,38 @@
 
 #include <numa.h>
 
+#include "velox/common/testutil/TestValue.h"
+
+using facebook::velox::common::testutil::TestValue;
+
 namespace facebook::velox::cxl {
 namespace {
+
+// Returns whether 'numaNode' has any CPUs attached.
+bool numaNodeHasCpus(int32_t numaNode) {
+  auto* cpus = numa_allocate_cpumask();
+  if (cpus == nullptr) {
+    return false;
+  }
+  const bool hasCpus =
+      numa_node_to_cpus(numaNode, cpus) == 0 && numa_bitmask_weight(cpus) > 0;
+  numa_free_cpumask(cpus);
+  return hasCpus;
+}
+
+// Verifies 'numaNode' is a CXL device. CXL expanders surface as CPU-less NUMA
+// nodes, so a node with CPUs is regular DRAM where allocations would silently
+// land off the CXL device.
+void checkNodeIsCxl(int32_t numaNode) {
+  bool hasCpus = numaNodeHasCpus(numaNode);
+  // Lets tests treat a DRAM node as CXL so the path runs on CXL-less CI hosts.
+  TestValue::adjust("facebook::velox::cxl::checkNodeIsCxl", &hasCpus);
+  VELOX_USER_CHECK(
+      !hasCpus,
+      "CXL memory allocator requires a CPU-less NUMA node; node has CPUs and "
+      "is likely DRAM: {}",
+      numaNode);
+}
 
 // Validates 'numaNode', returning it unchanged.
 int32_t validateNumaNode(int32_t numaNode) {
@@ -27,6 +57,7 @@ int32_t validateNumaNode(int32_t numaNode) {
   VELOX_CHECK_GE(numaNode, 0, "Invalid CXL NUMA node: {}", numaNode);
   VELOX_CHECK_LE(
       numaNode, numa_max_node(), "CXL NUMA node is out of range: {}", numaNode);
+  checkNodeIsCxl(numaNode);
   return numaNode;
 }
